Define function for rgb value generation and enable vivid button
Define a function to generate random rgb values and use this function to generate random rgb values for each cell in .cells class. Enable the vivid button so that once it is clicked, a mouseenter event is listened for which when detected calls the rgbGen function and changes the background color of the cell to the random rgb value generated by the function.